// src/beers/dto/create-beer.dto.ts
export interface CreateBeerDto {
    id: number;
    name: string;
    created: string;
    description: string;
    img_optimized: string;
    img_original: string;
    alcohol: number;
    beer_type: string;
    beer_type_family: string;
    beer_type_family_color: string;
    star_1: number;
    star_2: number;
    star_3: number;
    star_4: number;
    star_5: number;
    avg_rating: number;
    first_avg_rating: number;
    brewery: string;
    country: string;
}